+++ src/main/kotlin/us/frollo/frollosdk/network/api/MessagesAPI.kt
@@ -22,6 +22,7 @@ import retrofit2.http.GET
 import retrofit2.http.PUT
 import retrofit2.http.Path
 import retrofit2.http.QueryMap
+import us.frollo.frollosdk.model.api.messages.MessageBulkUpdateRequest
 import us.frollo.frollosdk.model.api.messages.MessageResponse
 import us.frollo.frollosdk.model.api.messages.MessageUpdateRequest
 import us.frollo.frollosdk.model.api.shared.PaginatedResponse
@@ -31,6 +32,7 @@ internal interface MessagesAPI {
         const val URL_UNREAD = "messages/unread"
         const val URL_MESSAGES = "messages"
         const val URL_MESSAGE = "messages/{message_id}"
+        const val URL_MESSAGES_BULK = "messages/bulk"
     }
 
     @GET(URL_MESSAGES)
@@ -44,4 +46,7 @@ internal interface MessagesAPI {
 
     @PUT(URL_MESSAGE)
     fun updateMessage(@Path("message_id") id: Long, @Body request: MessageUpdateRequest): Call<MessageResponse>
+
+    @PUT(URL_MESSAGES_BULK)
+    fun updateMessagesInBulk(@Body request: List<MessageBulkUpdateRequest>): Call<List<MessageResponse>>
 }